api/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
flask-migrate
flask-marshmallow
marshmallow-sqlalchemy
flask-jwt-extended
flask-mail
werkzeug<=0.16.1
psycopg2
selenium
numpy
pandas
requests
